fix: import model by folder on win
