+++ web/hooks/useImportModel.ts
@@ -66,43 +66,12 @@ const useImportModel = () => {
   const sanitizeFilePaths = useCallback(
     async (filePaths: string[]) => {
       if (!filePaths || filePaths.length === 0) return
-
-      const sanitizedFilePaths: FilePathWithSize[] = []
-      for (const filePath of filePaths) {
-        const fileStats = await fs.fileStat(filePath, true)
-        if (!fileStats) continue
-
-        if (!fileStats.isDirectory) {
-          const fileName = await baseName(filePath)
-          sanitizedFilePaths.push({
-            path: filePath,
-            name: fileName,
-            size: fileStats.size,
-          })
-        } else {
-          // allowing only one level of directory
-          const files = await fs.readdirSync(filePath)
-
-          for (const file of files) {
-            const fullPath = await joinPath([filePath, file])
-            const fileStats = await fs.fileStat(fullPath, true)
-            if (!fileStats || fileStats.isDirectory) continue
-
-            sanitizedFilePaths.push({
-              path: fullPath,
-              name: file,
-              size: fileStats.size,
-            })
-          }
-        }
+      const { unsupportedFiles, supportedFiles } = fs.getGgufFiles(
+        filePaths
+      ) as {
+        unsupportedFiles: FilePathWithSize[]
+        supportedFiles: FilePathWithSize[]
       }
-
-      const unsupportedFiles = sanitizedFilePaths.filter(
-        (file) => !file.path.endsWith('.gguf')
-      )
-      const supportedFiles = sanitizedFilePaths.filter((file) =>
-        file.path.endsWith('.gguf')
-      )
 
       const importingModels: ImportingModel[] = supportedFiles.map(
         ({ path, name, size }: FilePathWithSize) => ({
